Rename graphql_document crate to graphql_programs
