+++ server/graphql/programs/src/mutations/delete_document.rs
