+++ server/graphql/programs/src/mutations/encounter/insert.rs
@@ -6,7 +6,7 @@ use graphql_core::{
 use repository::{EncounterFilter, EqualFilter};
 use service::{
     auth::{Resource, ResourceAccessRequest},
-    document::encounter::{InsertEncounter, InsertEncounterError},
+    programs::encounter::{InsertEncounter, InsertEncounterError},
 };
 
 use crate::types::encounter::EncounterNode;
