+++ server/graphql/programs/src/mutations/insert_document_registry.rs
