+++ server/graphql/programs/src/queries/document_registry.rs
