+++ server/graphql/programs/src/queries/encounter.rs
