+++ server/graphql/programs/src/queries/json_schema.rs
